+++ src/components/retroHome/stripeBox/index.js
