+++ src/pages/index.js
@@ -1,6 +1,6 @@
 import React from "react";
 // import Link from 'gatsby-link'
-import RetroHome from "./retroHome";
+import RetroHome from "../components/retroHome";
 
 const IndexPage = () => (
   <RetroHome />
